requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
attrs==25.3.0
audioop-lts==0.2.1
build==1.2.2.post1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
discord.py==2.5.2
docopt==0.6.2
frozenlist==1.5.0
idna==3.10
multidict==6.3.2
packaging==24.2
pip-tools==7.4.1
pipreqs==0.4.13
propcache==0.3.1
pyproject_hooks==1.2.0
python-dotenv==1.1.0
requests==2.32.3
setuptools==78.1.0
urllib3==2.3.0
wheel==0.45.1
yarg==0.1.10
yarl==1.19.0
